requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
sqlalchemy
mysql-connector-python
alembic